requirements.txt
sentence-transformers
numpy
pytest
faiss-cpu